Move events, add editorconfig
