+++ api/src/main/java/com/owen1212055/biomevisuals/api/events/BiomeRegistrySendEvent.java
@@ -1,4 +1,4 @@
-package com.owen1212055.biomevisuals.api;
+package com.owen1212055.biomevisuals.api.events;
 
 import com.owen1212055.biomevisuals.api.types.biome.BiomeData;
 import org.bukkit.NamespacedKey;
@@ -19,10 +19,6 @@ public final class BiomeRegistrySendEvent extends RegistrySendEvent<BiomeData> {
 
     private final Map<NamespacedKey, BiomeData> registryEntries;
 
-    public static @NotNull HandlerList getHandlerList() {
-        return HANDLER_LIST;
-    }
-
     public BiomeRegistrySendEvent(final @NotNull Map<NamespacedKey, BiomeData> registryEntries) {
         this.registryEntries = registryEntries;
     }
@@ -37,25 +33,8 @@ public final class BiomeRegistrySendEvent extends RegistrySendEvent<BiomeData> {
         return HANDLER_LIST;
     }
 
-    @Override
-    public String toString() {
-        return "BiomeRegistrySendEvent{" +
-                "registryEntries=" + registryEntries +
-                ", isCancelled=" + isCancelled +
-                '}';
-    }
-
-    @Override
-    public boolean equals(Object o) {
-        if (this == o) return true;
-        if (o == null || getClass() != o.getClass()) return false;
-        BiomeRegistrySendEvent that = (BiomeRegistrySendEvent) o;
-        return isCancelled == that.isCancelled && registryEntries.equals(that.registryEntries);
-    }
-
-    @Override
-    public int hashCode() {
-        return Objects.hash(registryEntries, isCancelled);
+    public static @NotNull HandlerList getHandlerList() {
+        return HANDLER_LIST;
     }
 
 }
